Introduces a service (GenServer) to manage and load balance typesense nodes
In the event that there is no external load balancer wrapping requests
in front of Typesense SDKs are expected to distribute load against
instances.

This commit introduces a Supervisor/Genserver arrangement that keeps
track of a stack of retryable (healthy, maybe_healthy) nodes that
may be requested. next_node retrieves them in order and because this
service is distributed accros all users the load will be evenly
distributed across users.

Could do with one more refactoring pass.