Start stubbing in a request flow
This introduces a new Request module that will convert search params
into a call to a remote endpoint. By default, we use Telsa for this
but it is technically configurable. We also have a new Http module
that is a thin wrapper around tesla and makes calls at the boundary
mockable (via mox in this case).

After this, we can start building out all the endpoints in Typesense
that actually do the fetching/updating/deleting of content.