Allows Request to set node health
I'm not totally wild about Request.execute being so impure but it's
convenient to allow it to mark nodes as healthy/unhealthy and to
get the next node in the queue.

This is probably not the right/best/final design here - maybe this
request function really belongs in the client.

Collections (a new endppoint) can then do Client.request(...).