Support JSONL responses and plain text bodies (for import)
When importing documents, we actually need to send plain text to
Typesense. It knows how to speak JSONL. This commit extends Request.execute
to let it handle plain text bodies. AND imports return a plain text
response as well with success messages for each item that succeeded.
So, we attemp to roll through the messages and decode them as JSONL.

I might be convinced that we should just return a string. Decoding
is expensive and we may have many documents going through this
at one time.

Also, I think we are going to want to use Enum.stream functions
on import - we will certainly be getting very long lists of
documents.