Resolve timing issues
There is probably a better way to handle this but our set_unhealthy/2
function relies on Process.send_after to put a node back into rotation
after a given timeout. And because tests within the same module are
randomized, this means one set_unhealthy function call can alter
the outcome of a different test. So, as a stopgap, I have adjuseted
the timing of the set_unhealthy hooks to keep them from stepping
on each other.